+++ hadoop-ozone/common/src/main/java/org/apache/hadoop/ozone/om/multitenant/impl/BucketNameSpaceImpl.java
@@ -1,0 +1,69 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ *  with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.apache.hadoop.ozone.om.multitenant.impl;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.apache.hadoop.hdds.client.OzoneQuota;
+import org.apache.hadoop.hdds.fs.SpaceUsageSource;
+import org.apache.hadoop.ozone.om.multitenant.BucketNameSpace;
+import org.apache.hadoop.ozone.security.acl.OzoneObj;
+
+/**
+ * Implements BucketNameSpace.
+ */
+public class BucketNameSpaceImpl implements BucketNameSpace {
+  private final String bucketNameSpaceID;
+  private List<OzoneObj> bucketNameSpaceObjects;
+
+  public BucketNameSpaceImpl(String id) {
+    bucketNameSpaceID = id;
+    bucketNameSpaceObjects = new ArrayList<>();
+  }
+
+  @Override
+  public String getBucketNameSpaceID() {
+    return bucketNameSpaceID;
+  }
+
+  @Override
+  public List<OzoneObj> getBucketNameSpaceObjects() {
+    return bucketNameSpaceObjects;
+  }
+
+  @Override
+  public void addBucketNameSpaceObject(OzoneObj e) {
+    bucketNameSpaceObjects.add(e);
+  }
+
+  @Override
+  public SpaceUsageSource getSpaceUsage() {
+    return null;
+  }
+
+  @Override
+  public void setQuota(OzoneQuota quota) {
+
+  }
+
+  @Override
+  public OzoneQuota getQuota() {
+    return null;
+  }
+}
