+++ hadoop-ozone/common/src/main/java/org/apache/hadoop/ozone/om/protocolPB/OzoneManagerProtocolClientSideTranslatorPB.java
@@ -945,15 +945,15 @@ public final class OzoneManagerProtocolClientSideTranslatorPB
   @Override
   public TenantUserList listUsersInTenant(String tenantName, String prefix)
       throws IOException {
-   TenantListUserRequest request =
-       TenantListUserRequest.newBuilder().setTenantName(tenantName)
-           .setPrefix(prefix).build();
-   final  OMRequest omRequest = createOMRequest(Type.TenantListUser)
-       .setTenantListUserRequest(request).build();
-   final OMResponse response = submitRequest(omRequest);
-   final TenantListUserResponse resp =
-       handleError(response).getTenantListUserResponse();
-   return TenantUserList.fromProtobuf(resp.getTenantUserList());
+    TenantListUserRequest request =
+        TenantListUserRequest.newBuilder().setTenantName(tenantName)
+            .setPrefix(prefix).build();
+    final OMRequest omRequest = createOMRequest(Type.TenantListUser)
+        .setTenantListUserRequest(request).build();
+    final OMResponse response = submitRequest(omRequest);
+    final TenantListUserResponse resp =
+        handleError(response).getTenantListUserResponse();
+    return TenantUserList.fromProtobuf(resp.getTenantUserList());
   }
 
   @Override
